+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.3.0] - 2023-01-23
+- Added: encore contracts support ([#16])
+
+## [1.2.0] - 2023-01-18
+- Added: ExternalElementInterface ([#15])
+- Changed: [Attention!] added data attributes variable to wrapper elements. Please adjust your custom templates. ([#15])
+- Changed: refactored js code ([#15])
+- Fixed: compatibility issues with contao 4.13 ([#15])
+- Fixed: palette issues in tl_page and tl_content ([#15])
+- Fixed: warning with php 8 ([#15])
+
 ## [1.1.1] - 2022-09-16
 - Added: iframe aria-labels
 
@@ -194,3 +205,6 @@ All notable changes to this project will be documented in this file.
 
 Initial version
 
+
+[#16]: https://github.com/heimrichhannot/contao-video-bundle/pull/16
+[#15]: https://github.com/heimrichhannot/contao-video-bundle/pull/15